# feat: Complete Stage 2-4 Implementation with Mobile Optimization & Deployment