--- a/src/stage4/__init__.py
+++ b/src/stage4/__init__.py
@@ -1,0 +1,23 @@
+"""
+Stage 4: Cascade Detection System Integration and Mobile Optimization
+====================================================================
+
+This module implements the fourth stage of the cascade deepfake detection system,
+focusing on unified cascade integration and mobile deployment optimization.
+
+Key Components:
+- Unified cascade detection system (Stage 1 → Stage 2 → Stage 3)
+- Dynamic threshold strategies with temperature scaling
+- Mobile optimization through QAT + Knowledge Distillation
+- Video processing and batch inference capabilities
+- Performance benchmarking and deployment testing
+
+Target Performance:
+- <200ms per frame on mobile hardware
+- >0.95 overall system AUC
+- <500MB total memory footprint
+- >75% model size reduction
+"""
+
+__version__ = "1.0.0"
+__author__ = "AWARE-NET Research Team"