--- a/src/stage4/mobile_deployment/__init__.py
+++ b/src/stage4/mobile_deployment/__init__.py
@@ -1,0 +1,19 @@
+"""
+Mobile Deployment Utilities
+==========================
+
+Utilities for exporting optimized models to mobile deployment formats
+including ONNX and TensorFlow Lite conversion.
+"""
+
+from .onnx_exporter import ONNXExporter
+from .tflite_converter import TFLiteConverter
+from .mobile_inference import MobileInferenceWrapper
+from .deployment_validator import DeploymentValidator
+
+__all__ = [
+    'ONNXExporter',
+    'TFLiteConverter', 
+    'MobileInferenceWrapper',
+    'DeploymentValidator'
+]