… .claude/settings.local.json
… .ipynb_checkpoints/README-checkpoint.md
… README.md
… project_instruction/progress_record/20250725.md
… project_instruction/progress_record/20250725_final.md
… src/stage2/evaluate_stage2.py
… src/stage2/feature_extractor.py
… src/stage3/create_meta_dataset.py
… src/stage3/train_meta_model.py
… src/stage4/README.md
… src/stage4/TESTING.md
… src/stage4/benchmark_cascade.py
… src/stage4/cascade_detector.py
… src/stage4/optimize_for_mobile.py
… src/stage4/run_mobile_optimization.sh
… src/stage4/test_mobile_optimization.py
… src/stage4/test_stage4_integration.py
… src/stage4/video_processor.py